+++ CHANGELOG.md
@@ -2,11 +2,26 @@
 
 ## [Unreleased]
 
+- Remove unused pystac.validation import ([#1583](https://github.com/stac-utils/pystac/pull/1583))
+
+## [v1.14.1] - 2025-09-18
+
+### Fixed
+
+- Support older versions of urllib3 ([#1580](https://github.com/stac-utils/pystac/pull/1580))
+
+## [v1.14.0] - 2025-09-11
+
 ### Added
 
 - Added to pystac.MediaType values VND_APACHE_PARQUET and VND_ZARR with the current standard
   media type value for these types and new media types COPC and VND_PMTILES
   ([#1554](https://github.com/stac-utils/pystac/pull/1554))
+
+### Changed
+
+- Pass "User-Agent": "pystac/3.13.0" in request headers to work around urllib's default "User-Agent" being blocked by readthedocs ([#1576](https://github.com/stac-utils/pystac/pull/1576))
+- Include a plain text repr in notebooks so untrusted notebooks look better ([#1577](https://github.com/stac-utils/pystac/pull/1577))
 
 ### Fixed
 
@@ -942,7 +957,9 @@ use `Band.create`
 
 Initial release.
 
-[Unreleased]: <https://github.com/stac-utils/pystac/compare/v1.13.0..main>
+[Unreleased]: <https://github.com/stac-utils/pystac/compare/v1.14.1..main>
+[v1.14.1]: <https://github.com/stac-utils/pystac/compare/v1.14.0..v1.14.1>
+[v1.14.0]: <https://github.com/stac-utils/pystac/compare/v1.13.0..v1.14.0>
 [v1.13.0]: <https://github.com/stac-utils/pystac/compare/v1.12.2..v1.13.0>
 [v1.12.2]: <https://github.com/stac-utils/pystac/compare/v1.12.1..v1.12.2>
 [v1.12.1]: <https://github.com/stac-utils/pystac/compare/v1.12.0..v1.12.1>